# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for SRN (Skill Requirement Network)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Retrieve chat history
      parameters:
        - name: senderId
          in: query
          required: true
          schema:
            type: string
        - name: receiverId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessage"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        receiverId:
          type: string
        text:
          type: string
        attachmentUrl:
          type: string
        attachmentName:
          type: string
        attachmentSize:
          type: string
        quoteId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - senderId
        - receiverId
        - text
    CreateMessage:
      type: object
      properties:
        senderId:
          type: string
        receiverId:
          type: string
        text:
          type: string
        attachmentUrl:
          type: string
        attachmentName:
          type: string
        attachmentSize:
          type: string
        quoteId:
          type: string
      required:
        - senderId
        - receiverId
        - text
...